Allow no-op usage with rspec 3.
