+++ spec/gemfiles/rspec-2
@@ -1,5 +1,5 @@
 source 'https://rubygems.org'
 
-gem 'rspec'
+gem 'rspec', '~> 2.11.0'
 
 gemspec :path => '../../'
