--- spec/gemfiles/rspec-master
@@ -1,6 +1,0 @@
-source 'https://rubygems.org'
-
-gem 'rspec'
-gem 'rspec-mocks', :git => 'git://github.com/rspec/rspec-mocks.git'
-
-gemspec :path => '../../'
